Modificaciones y Actualizaciones varias

+++ delete_note.php
@@ -1,18 +1,18 @@
 <?php
-
+session_start();
 include("db.php");
+if (isset($_SESSION['user'])){
+  if(isset($_GET['id'])) {
+    $id = $_GET['id'];
+    $query = "DELETE FROM notes WHERE id = $id";
+    $result = mysqli_query($conn, $query);
+    if(!$result) {
+      die("Query Failed.");
+    }
 
-if(isset($_GET['id'])) {
-  $id = $_GET['id'];
-  $query = "DELETE FROM notes WHERE id = $id";
-  $result = mysqli_query($conn, $query);
-  if(!$result) {
-    die("Query Failed.");
+    $_SESSION['message'] = 'Nota Borrada';
+    $_SESSION['message_type'] = 'danger';
+    header('Location: papelera.php');
   }
-
-  $_SESSION['message1'] = 'Note Removed Successfully';
-  $_SESSION['message_type'] = 'danger';
-  header('Location: main.php');
 }
-
 ?>
